// src/payload-types.ts
export interface Media {
  id: string
  alt?: string | null
  caption?: string | null
  url?: string | null
  thumbnailURL?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
  focalX?: number | null
  focalY?: number | null
  updatedAt: string
  createdAt: string
}